Add test for default to get and port to pytest
