+++ docs/maintaining/index.md
@@ -8,4 +8,5 @@
    overview
    supported-python-versions
    making-a-release
+   testing
 ```
