+++ opsdroid/conftest.py
@@ -1,15 +1,6 @@
 """Pytest config for all opsdroid tests."""
-import pytest
+from opsdroid.testing import opsdroid
 
-from opsdroid.core import OpsDroid
+from opsdroid.cli.start import configure_lang
 
-
-@pytest.fixture
-def opsdroid():
-    """Fixture with a plain instance of opsdroid.
-
-    Will yield an instance of opsdroid which hasn't been loaded.
-
-    """
-    with OpsDroid(config={}) as opsdroid:
-        yield opsdroid
+configure_lang({})
