+++ requirements_test.txt
@@ -16,3 +16,4 @@ sphinx==3.0.4
 sphinx-click==2.5.0
 recommonmark==0.6.0
 deadlinks==0.3.2
+mock
