+++ setup.cfg
@@ -4,6 +4,7 @@ universal = 1
 [tool:pytest]
 testpaths = opsdroid tests
 norecursedirs = .git testing_config
+mock_use_standalone_module = true
 
 [flake8]
 max-line-length = 80
